requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
cssselect==1.3.0
distro==1.9.0
fake-useragent==2.2.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-core==0.3.56
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langsmith==0.3.37
lxml==5.4.0
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
mdurl==0.1.2
openai==1.76.0
orjson==3.10.16
outcome==1.3.0.post0
packaging==24.2
parse==1.20.2
pydantic==2.11.3
pydantic_core==2.33.1
pyee==11.1.1
Pygments==2.19.1
PyMuPDF==1.25.5
pymupdf4llm==0.0.21
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-html==0.10.0
requests-toolbelt==1.0.0
rich==14.0.0
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.40
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.20
w3lib==2.3.1
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.21.0
zstandard==0.23.0
